The issues with the type checker should be fixed.
The problem is that the type constraint system reducer only reduces the system such that a solution is easily attainable, but it isn't always reduced to a form where there aren't free variables. As a result, getting the type of a variable from a substitution can take more than one iteration.